+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # [next]
 - BREAKING CHANGE:
   - remove `BonfireTiledWidget`. now use `BonfireWidget` pass map `MapWordByTiled`
+- improvements in `simpleAttackMeleeByDirection` and `simpleAttackMelee`. now it's not necessary set animation to all directions, only to right.
 
 # [2.8.1]
 - Adds `onContactExit` in `Sensor` mixin.